Modified safety model for pulled-in dependencies
Dependencies for your own defined functions are assumed to be safe now. This means if you have a dependency that deletes your system, it is allowed (it would only botch the OpenFaaS instance anyways). This simplifies usage and development. Note this does not apply to calls, which have the original model